pyproject.toml
[project]
name = "benchmark-dataloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "av>=14.3.0",
    "bitstring>=4.3.1",
    "dotenv>=0.9.9",
    "ebmlite>=3.3.1",
    "fastparquet>=2024.11.0",
    "filetype>=1.2.0",
    "jsonargparse>=4.38.0",
    "kaggle>=1.7.4.2",
    "kornia>=0.8.0",
    "megatron-energon[s3]>=6.0.1",
    "mosaicml-streaming>=0.12.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pytorch-lightning>=2.5.1",
    "pyyaml>=6.0.2",
    "sortedcontainers>=2.4.0",
    "soundfile>=0.13.1",
    "torchvision>=0.21.0",
    "webdataset>=0.2.111",
]
